keeper/genesis.go
package keeper

import (
	"context"

	examplev1 "github.com/cosmosregistry/example/api/v1"
)

// InitGenesis initializes the module state from a genesis state.
func (k *Keeper) InitGenesis(ctx context.Context, data *examplev1.GenesisState) error {
	if err := k.Params.Set(ctx, *data.Params); err != nil { // TODO: not good.
		return err
	}

	for _, counter := range data.Counters {
		if err := k.Counter.Set(ctx, counter.Address, counter.Count); err != nil {
			return err
		}
	}

	return nil
}

// ExportGenesis exports the module state to a genesis state.
func (k *Keeper) ExportGenesis(ctx context.Context) (*examplev1.GenesisState, error) {
	params, err := k.Params.Get(ctx)
	if err != nil {
		return nil, err
	}

	var counters []*examplev1.Counter
	if err := k.Counter.Walk(ctx, nil, func(address string, count uint64) (bool, error) {
		counters = append(counters, &examplev1.Counter{
			Address: address,
			Count:   count,
		})

		return false, nil
	}); err != nil {
		return nil, err
	}

	return &examplev1.GenesisState{
		Params:   &params,
		Counters: counters,
	}, nil
}
